Add SPCT_UNIX for MacOSX in spctOS
SPCT_UNIX flag was disabled on MacOSX in order to prevent the generation of the
install target. Speect plugins are searched in the installation directory and,
if not found, in the build directory. CMake clear RPATH in dynamic libraries
(plugins included) during the installation phase. If the installation path
will results in libspct library being installed in unusual directory (i.e.: one
that is not in the dynamic loader search path), this will result in issues
when loading plugins as libspct will not be found.
A solution is to set DYLD_LIBRARY_PATH to the contain the libspct directory.
Another solution is to only install libraries in system directories.
Another possibility is to avoid install target completely, unless it is possible
to apply one of the above solution.
On some systems (e.g.: Linux) the problem will not be noticed as the libspct
is loaded by the program and the linker is smart enough to detect it. On Linux
DYLD_LIBRARY_PATH equivalent is called LD_LIBRARY_PATH.